+++ udata_front/theme/gouvfr/assets/js/types.ts
@@ -32,6 +32,8 @@ export type PublishingFormAccordionState = AccordionFunctionalState | DSFRInfoSt
 
 export type AccordionState = DSFRFormDefaultState | AccordionFunctionalState | DSFRInfoState;
 
+export type AdminBadgeState = DSFRFormDefaultState | FormFunctionalState | DSFRInfoState;
+
 export type ClosedFormats = typeof CLOSED_FORMATS[number];
 
 export type DatasetRemoteFile = { description?: string; filetype: RemoteResourceFileType; format: string; mime: string; schema?: string; title: string; type: ResourceType; url: string; };
@@ -158,6 +160,10 @@ export type Me = User & {
 }
 
 export type AxisAlignment = "start" | "center" | "end";
+
+export type SortDirection = 'asc' | 'desc';
+
+export type DatasetSortedBy = 'title' | 'created' | 'last_update' | 'reuses' | 'followers' | 'views';
 
 export type SpatialZone = {
   code: string;