+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Current (in progress)
 
-- Nothing yet
+- Add index page with search for dataservices [#516](https://github.com/datagouv/udata-front/pull/516)
+- Add beta admin datasets page [#371](https://github.com/etalab/udata-front/pull/371)
 
 ## 5.2.0 (2024-09-13)